Basic Resource Tick
Implemented a very crude method in BuildingManager which totals up all of the generated, demanded, and available resources on each Tick (one second).

Buildings will attempt to satisfy their own demand first through on-site storage (if possible), and will then attempt to satisfy demand via other buildings if it was not adequate.

Note: This does not attempt to deal with the tank/warehouse buildings at this time. (Resources are not stored or withdrawn)

Buildings that are unable to satisfy their required intake then have their output reduced proportionally to the amount of demand left unsatisfied. Example: Coal plant needs 2 coal, only is able to satisfy with 1 unit of coal; its' ActualResourceGeneratedPerSec is 0.5 units of power.